Added talisman ignores.
