+++ .talismanrc
@@ -1,3 +1,5 @@
 fileignoreconfig:
 - filename: db/schema.rb
   checksum: b9812e78c2922e34d8894a76a3af7120146f8e9bda2eaf9d53cdc2adbf4295c5
+- filename: app/resources/v1/pacbio/run_resource.rb
+  checksum: 933961527fa2cf90add25e17e842c31463d80d3c683c26a44206790575968619
