--- app/resources/v1/pacbio/plate_resource.rb
@@ -1,14 +1,0 @@
-# frozen_string_literal: true
-
-module V1
-  module Pacbio
-    # PlateResource
-    class PlateResource < JSONAPI::Resource
-      model_name 'Pacbio::Plate'
-
-      attributes :pacbio_run_id
-
-      has_many :wells, class_name: 'Well'
-    end
-  end
-end
