+++ app/resources/v1/pacbio/runs/well_library_resource.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module V1
+  module Pacbio
+    module Runs
+      # WellLibraryResource
+      class WellLibraryResource < JSONAPI::Resource
+        model_name 'Pacbio::WellLibrary'
+
+        attributes :barcode
+      end
+    end
+  end
+end
