+++ spec/requests/v1/pacbio/runs/plates_spec.rb
@@ -10,14 +10,14 @@ RSpec.describe 'PlatesController', type: :request do
     let!(:wells) {create_list(:pacbio_well, 5, plate: plate2)}
 
     it 'returns a list of plates' do
-      get v1_pacbio_plates_path, headers: json_api_headers
+      get v1_pacbio_runs_plates_path, headers: json_api_headers
       expect(response).to have_http_status(:success)
       json = ActiveSupport::JSON.decode(response.body)
       expect(json['data'].length).to eq(2)
     end
 
     it 'returns the correct attributes' do
-      get "#{v1_pacbio_plates_path}?include=wells", headers: json_api_headers
+      get "#{v1_pacbio_runs_plates_path}?include=wells", headers: json_api_headers
 
       expect(response).to have_http_status(:success)
       json = ActiveSupport::JSON.decode(response.body)
@@ -51,16 +51,16 @@ RSpec.describe 'PlatesController', type: :request do
       end
 
       it 'has a ok status' do
-        post v1_pacbio_plates_path, params: body, headers: json_api_headers
+        post v1_pacbio_runs_plates_path, params: body, headers: json_api_headers
         expect(response).to have_http_status(:created)
       end
 
       it 'creates a plate' do
-        expect { post v1_pacbio_plates_path, params: body, headers: json_api_headers }.to change { Pacbio::Plate.count }.by(1)
+        expect { post v1_pacbio_runs_plates_path, params: body, headers: json_api_headers }.to change { Pacbio::Plate.count }.by(1)
       end
 
       it 'has the correct attributes' do
-        post v1_pacbio_plates_path, params: body, headers: json_api_headers
+        post v1_pacbio_runs_plates_path, params: body, headers: json_api_headers
         plate = Pacbio::Plate.first
         expect(plate.run).to eq run
       end
@@ -80,16 +80,16 @@ RSpec.describe 'PlatesController', type: :request do
       end
 
       it 'has a ok unprocessable_entity' do
-        post v1_pacbio_plates_path, params: body, headers: json_api_headers
+        post v1_pacbio_runs_plates_path, params: body, headers: json_api_headers
         expect(response).to have_http_status(:unprocessable_entity)
       end
 
       it 'does not create a plate' do
-        expect { post v1_pacbio_plates_path, params: body, headers: json_api_headers }.to_not change(Pacbio::Plate, :count)
+        expect { post v1_pacbio_runs_plates_path, params: body, headers: json_api_headers }.to_not change(Pacbio::Plate, :count)
       end
 
       it 'has an error message' do
-        post v1_pacbio_plates_path, params: body, headers: json_api_headers
+        post v1_pacbio_runs_plates_path, params: body, headers: json_api_headers
         json = ActiveSupport::JSON.decode(response.body)
         errors = json['data']['errors']
         expect(errors['run']).to be_present
@@ -116,18 +116,18 @@ RSpec.describe 'PlatesController', type: :request do
       end
 
       it 'has a ok status' do
-        patch v1_pacbio_plate_path(plate), params: body, headers: json_api_headers
+        patch v1_pacbio_runs_plate_path(plate), params: body, headers: json_api_headers
         expect(response).to have_http_status(:ok)
       end
 
       it 'updates a plate' do
-        patch v1_pacbio_plate_path(plate), params: body, headers: json_api_headers
+        patch v1_pacbio_runs_plate_path(plate), params: body, headers: json_api_headers
         plate.reload
         expect(plate.run).to eq new_run
       end
 
       it 'returns the correct attributes' do
-        patch v1_pacbio_plate_path(plate), params: body, headers: json_api_headers
+        patch v1_pacbio_runs_plate_path(plate), params: body, headers: json_api_headers
         json = ActiveSupport::JSON.decode(response.body)
         expect(json['data']['id']).to eq plate.id.to_s
       end
@@ -147,12 +147,12 @@ RSpec.describe 'PlatesController', type: :request do
       end
 
       it 'has a ok unprocessable_entity' do
-        patch v1_pacbio_plate_path(123), params: body, headers: json_api_headers
+        patch v1_pacbio_runs_plate_path(123), params: body, headers: json_api_headers
         expect(response).to have_http_status(:unprocessable_entity)
       end
 
       it 'has an error message' do
-        patch v1_pacbio_plate_path(123), params: body, headers: json_api_headers
+        patch v1_pacbio_runs_plate_path(123), params: body, headers: json_api_headers
         expect(JSON.parse(response.body)["data"]).to include("errors" => "Couldn't find Pacbio::Plate with 'id'=123")
       end
     end
@@ -164,28 +164,28 @@ RSpec.describe 'PlatesController', type: :request do
 
     context 'on success' do
       it 'has a status of ok' do
-        delete v1_pacbio_plate_path(plate), headers: json_api_headers
+        delete v1_pacbio_runs_plate_path(plate), headers: json_api_headers
         expect(response).to have_http_status(:no_content)
       end
 
       it 'deletes the plate' do
-        expect { delete v1_pacbio_plate_path(plate), headers: json_api_headers }.to change { Pacbio::Plate.count }.by(-1)
+        expect { delete v1_pacbio_runs_plate_path(plate), headers: json_api_headers }.to change { Pacbio::Plate.count }.by(-1)
       end
 
       it 'deletes the wells' do
-        expect { delete v1_pacbio_plate_path(plate), headers: json_api_headers }.to change { Pacbio::Well.count }.by(-1)
+        expect { delete v1_pacbio_runs_plate_path(plate), headers: json_api_headers }.to change { Pacbio::Well.count }.by(-1)
       end
     end
 
     context 'on failure' do
 
       it 'does not delete the plate' do
-        delete "/v1/pacbio/plates/123", headers: json_api_headers
+        delete "/v1/pacbio/runs/plates/123", headers: json_api_headers
         expect(response).to have_http_status(:unprocessable_entity)
       end
 
       it 'has an error message' do
-        delete "/v1/pacbio/plates/123", headers: json_api_headers
+        delete "/v1/pacbio/runs/plates/123", headers: json_api_headers
         data = JSON.parse(response.body)['data']
         expect(data['errors']).to be_present
       end
